Starting day11 with green acceptance test and dummy implementation
are 7:30 am @ Pinzolo and today it's the Polenta & Deploy day !
i am near the bar of the Ferrari hotel that is opening the breakfast
and i am  waiting @bombo82 and @marcobacis to join me for this kata

the idea was to start a new day together (instead of complete the day09),
in order to start with the same knowledge of the problem and
i choose the day 11 but ..... nobody comes!

they are still sleeping? they are in the shower? who knows..