Day 17 second part solved ! :tada:
Duplicated main second part function
LowestRegisterValueToPrintOutTheProgramItself to create another version
of it (LowestRegisterValueToPrintOutThe3BitConsumingLoopProgramItself)
that use 3-bit-consuming optimization to quickly find the register A candidate.

The solution is valid only for programs that consume 3-bit-at-time the
register A value with an output loop (like the file and the 2nd provided example).